+++ packages/abstract-document/src/abstract-document-exporters/__tests__/pdf/test-defs/simple-table-center-alignment.tsx
@@ -38,76 +38,76 @@ export const test: ExportTestDef = {
   ),
   expectedPdfJson: {
     formImage: {
+      Transcoder: "pdf2json@1.2.3 [https://github.com/modesty/pdf2json]",
       Agency: "",
       Id: {
         AgencyId: "",
+        Name: "",
         MC: false,
         Max: 1,
-        Name: "",
         Parent: "",
       },
       Pages: [
         {
-          Boxsets: [],
-          Fields: [],
-          Fills: [
-            {
-              clr: 1,
-              h: 0,
-              w: 0,
-              x: 0,
-              y: 0,
-            },
-          ],
+          Height: 52.625,
           HLines: [
             {
               x: 6.094,
-              y: 0,
-              w: 1.5,
+              y: 0.679,
+              w: 3,
               l: 25,
             },
             {
-              x: 6.094,
-              y: 0.723,
-              w: 1.5,
-              l: 25,
-            },
-          ],
-          Height: 52.625,
-          Texts: [
-            {
-              A: "left",
-              R: [
-                {
-                  S: -1,
-                  T: "CenterAlignedTable",
-                  TS: [0, 13, 0, 0],
-                },
-              ],
-              clr: 0,
-              sw: 0.32553125,
-              w: 88.37,
-              x: 5.844,
-              y: -0.301,
+              x: 6.075,
+              y: 0.044,
+              w: 3,
+              l: 25.038,
             },
           ],
           VLines: [
             {
-              x: 6.094,
+              x: 31.05,
               y: 0,
-              w: 1.5,
-              l: 0.723,
+              w: 3,
+              l: 0.741,
             },
             {
-              x: 31.094,
+              x: 6.138,
               y: 0,
-              w: 1.5,
-              l: 0.723,
+              w: 3,
+              l: 0.741,
             },
           ],
+          Fills: [
+            {
+              x: 0,
+              y: 0,
+              w: 0,
+              h: 0,
+              clr: 1,
+            },
+          ],
+          Texts: [
+            {
+              x: 5.844,
+              y: -0.301,
+              w: 88.37,
+              sw: 0.32553125,
+              clr: 0,
+              A: "left",
+              R: [
+                {
+                  T: "CenterAlignedTable",
+                  S: -1,
+                  TS: [0, 13, 0, 0],
+                },
+              ],
+            },
+          ],
+          Fields: [],
+          Boxsets: [],
         },
       ],
-      Transcoder: "pdf2json@1.2.3 [https://github.com/modesty/pdf2json]",
       Width: 37.188,
     },
   },