+++ packages/abstract-document/src/abstract-document-exporters/__tests__/pdf/test-defs/single-textrun-with-justify-alignment.tsx
@@ -1,0 +1,66 @@
+import React from "react";
+import { ExportTestDef } from "../export-test-def";
+import * as AD from "../../../../index";
+import { Paragraph, AbstractDoc, Section, TextRun } from "../../../../abstract-document-jsx";
+
+export const test: ExportTestDef = {
+  name: "Single textrun with Justify alignment",
+  abstractDocJsx: (
+    <AbstractDoc>
+      <Section>
+        <Paragraph style={AD.AbstractDoc.ParagraphStyle.create({ alignment: "Justify" })}>
+          <TextRun text="This is a text" />
+        </Paragraph>
+      </Section>
+    </AbstractDoc>
+  ),
+  expectedPdfJson: {
+    formImage: {
+      Transcoder: "pdf2json@1.2.3 [https://github.com/modesty/pdf2json]",
+      Agency: "",
+      Id: {
+        AgencyId: "",
+        Name: "",
+        MC: false,
+        Max: 1,
+        Parent: "",
+      },
+      Pages: [
+        {
+          Height: 52.625,
+          HLines: [],
+          VLines: [],
+          Fills: [
+            {
+              x: 0,
+              y: 0,
+              w: 0,
+              h: 0,
+              clr: 1,
+            },
+          ],
+          Texts: [
+            {
+              x: -0.25,
+              y: -0.301,
+              w: 56.13,
+              sw: 0.32553125,
+              clr: 0,
+              A: "left",
+              R: [
+                {
+                  T: "This%20is%20a%20text",
+                  S: -1,
+                  TS: [0, 13, 0, 0],
+                },
+              ],
+            },
+          ],
+          Fields: [],
+          Boxsets: [],
+        },
+      ],
+      Width: 37.188,
+    },
+  },
+};